konvert_app/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
python-jose
passlib==1.7.4
bcrypt==4.0.1
boto3
pillow
python-docx
python-dotenv
email-validator
jinja2
